requirements.txt
# Core dependencies
python-telegram-bot==20.7
sqlalchemy==2.0.25
httpx==0.25.2
fastapi==0.109.0
uvicorn[standard]==0.26.0
python-dotenv==1.0.0

# Scheduling
apscheduler==3.10.4

# WebSocket support
websockets==12.0

# Additional utilities
python-multipart==0.0.6
pydantic==2.5.3
